requirements.txt
streamlit==1.25.0
transformers==4.34.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.101.0
